+++ src/pages/api/download.ts
@@ -1,7 +1,7 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next';
-import Kite from '../../common/kite';
-import ReadableString from '../../common/utilities';
+import Kite from '@/common/kite';
+import ReadableString from '@/common/utilities';
 type File = string;
 
 export default async function handler(
