+++ src/pages/api/kite/create.ts
@@ -1,7 +1,7 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next';
-import Kite from '../../common/kite';
-import { Kafka } from 'kafkajs';
+import Kite from '@/common/kite';
+// import { Kafka } from 'kafkajs';
 
 type Data = {
   reply?: string;
