+++ src/pages/api/kite/getConfigFile.ts
@@ -1,6 +1,6 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next';
-import Kite from '../../common/kite';
+import Kite from '@/common/kite';
 
 type ConfigFile = KiteConfigFile;
