+++ src/pages/api/kite/getSetup.ts
@@ -1,13 +1,13 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next';
-import Kite from '../../common/kite';
+import Kite from '@/common/kite';
 
-type Config = KiteConfig;
+type Setup = KiteSetup;
 
 export default async function handler(
   req: NextApiRequest,
-  res: NextApiResponse<Config>
+  res: NextApiResponse<Setup>
 ) {
-  const config = await Kite.getConfig();
-  res.status(200).json(config);
+  const setup = await Kite.getSetup();
+  res.status(200).json(setup);
 }
